Refactor test mocks into shared mock_repositories file
Moved all mock and fake repository classes into a new shared file at test/mocks/mock_repositories.dart. Updated bloc test files to import and use these shared mocks, reducing duplication and improving maintainability.
